+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] / 2026-04-06
+### Features
+- Support `Revit 2027` and `Revit Preview`.
+### Updated
+- Update `IsInAddInContext` to support `Revit 2027`, change to `InAddInEventContext`. (Fix #39)
+
 ## [0.8.0] / 2024-12-27 - 2025-02-18
 ### Features
 - Support `RevitApplication` static class.
@@ -389,6 +395,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[0.9.0]: ../../compare/0.8.0...0.9.0
 [0.8.0]: ../../compare/0.7.1...0.8.0
 [0.7.1]: ../../compare/0.7.0...0.7.1
 [0.7.0]: ../../compare/0.6.2...0.7.0